added a few missing speeds for edge costing

+++ src/sif/motorcyclecost.cc
@@ -210,7 +210,7 @@ public:
    * @param   speed A speed for a road segment/edge.
    * @return  Returns the cost and time (seconds)
    */
-  virtual Cost EdgeCost(const baldr::DirectedEdge* edge, const float speed) const;
+  virtual Cost EdgeCost(const baldr::DirectedEdge* edge, const uint64_t speed) const;
 
   /**
    * Returns the cost to make the transition from the predecessor edge.
@@ -494,7 +494,7 @@ bool MotorcycleCost::Allowed(const baldr::NodeInfo* node) const {
   return (node->access() & kMotorcycleAccess);
 }
 
-Cost MotorcycleCost::EdgeCost(const baldr::DirectedEdge* edge, const float speed) const {
+Cost MotorcycleCost::EdgeCost(const baldr::DirectedEdge* edge, const uint64_t speed) const {
 
   if (edge->use() == Use::kFerry) {
     float sec = (edge->length() * speedfactor_[edge->speed()]);
